finish abilities

+++ src/main/java/com/example/service/implementations/AbilityServiceImpl.java
@@ -4,6 +4,7 @@ import com.example.dto.PivotDTO.AbilityDTO;
 import com.example.entity.enums.AbilityType;
 import com.example.entity.pivots.characters.Ability;
 import com.example.exceptions.AbilityAlreadyExistException;
+import com.example.exceptions.AbilityNotFoundException;
 import com.example.repository.AbilityRepository;
 import com.example.service.interfaces.AbilityService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,8 +25,8 @@ public class AbilityServiceImpl implements AbilityService {
 
     @Override
     public AbilityDTO addAbility(AbilityDTO abilityDTO) throws AbilityAlreadyExistException {
-        Ability ability = abilityRepository.findById(abilityDTO.getId()).orElse(null);
-        if(ability != null){
+        Ability ability = abilityRepository.findByDescription(abilityDTO.getDescription()).orElse(null);
+        if (ability != null) {
             throw new AbilityAlreadyExistException("Ability already exists");
         }
         ability = abilityRepository.save(abilityDTO.toAbility());
@@ -41,10 +42,10 @@ public class AbilityServiceImpl implements AbilityService {
     }
 
     @Override
-    public AbilityDTO findById(Long abilityId) {
+    public AbilityDTO findById(Long abilityId) throws AbilityNotFoundException {
         Ability ability = abilityRepository.findById(abilityId).orElse(null);
         if(ability == null){
-            throw new NullPointerException("Ability with id" + abilityId + " not found");
+            throw new AbilityNotFoundException("Ability with id: " + abilityId + " not found");
         }
         return AbilityDTO.fromAbility(ability);
     }
@@ -58,13 +59,19 @@ public class AbilityServiceImpl implements AbilityService {
     }
 
     @Override
-    public boolean delete(Long abilityId) {
+    public void update(Long abilityId, AbilityDTO updatedAbility) throws AbilityNotFoundException {
+        AbilityDTO abilityDTO = this.findById(abilityId);
+        abilityDTO.checkFields(updatedAbility);
+        abilityRepository.setAbilityInfoById(abilityId, abilityDTO.getType(), abilityDTO.getDescription());
+    }
+
+    @Override
+    public void delete(Long abilityId) throws AbilityNotFoundException {
         Ability ability = abilityRepository.findById(abilityId).orElse(null);
-        if(ability == null){
-            throw new NullPointerException("Ability with id" + abilityId + " not found");
+        if(ability == null || ability.getId() == null){
+            throw new AbilityNotFoundException("Ability with id: " + abilityId + " not found");
         }
         abilityRepository.deleteById(abilityId);
-        ability = abilityRepository.findById(abilityId).orElse(null);
-        return ability == null;
+        abilityRepository.deleteFromCharacterTable(abilityId);
     }
 }
